fix(app): server error not showing properly
