+++ apps/app/src/app/[locale]/address/[id]/page.tsx
@@ -1,5 +1,4 @@
 import { Suspense } from 'react';
-import { ErrorBoundary } from 'react-error-boundary';
 
 import AccountTabs from '@/components/app/Address/AccountTabs';
 import Balance from '@/components/app/Address/Balance';
@@ -41,11 +40,9 @@ export default async function AddressIndex(props: {
 
       <div className="py-2"></div>
 
-      <ErrorBoundary fallback={<TabSkeletion error reset />}>
-        <Suspense fallback={<TabSkeletion />} key={JSON.stringify(rest)}>
-          <AccountTabs id={id} parse={parse} searchParams={searchParams} />
-        </Suspense>
-      </ErrorBoundary>
+      <Suspense fallback={<TabSkeletion />} key={JSON.stringify(rest)}>
+        <AccountTabs id={id} parse={parse} searchParams={searchParams} />
+      </Suspense>
     </>
   );
 }
