+++ apps/app/src/components/app/Address/NFTTransactions.tsx
@@ -11,6 +11,10 @@ const NFTTransactions = async ({ id, searchParams }: any) => {
     getRequest(`account/${id}/nft-txns/count`, searchParams, options),
   ]);
 
+  if (data.message === 'Error') {
+    throw new Error(`Server Error : ${data.error}`);
+  }
+
   return (
     <>
       <NFTTransactionActions
