+++ apps/app/src/components/app/Address/TokenTransactions.tsx
@@ -11,6 +11,10 @@ const TokenTransactions = async ({ id, searchParams }: any) => {
     getRequest(`account/${id}/ft-txns/count`, searchParams, options),
   ]);
 
+  if (data.message === 'Error') {
+    throw new Error(`Server Error : ${data.error}`);
+  }
+
   return (
     <TokenTxnsActions
       count={count?.txns?.[0]?.count}
