+++ apps/app/src/components/app/Blocks/List.tsx
@@ -5,6 +5,10 @@ import ListActions from './ListActions';
 const List = async ({ cursor }: { cursor: string }) => {
   const data = await getRequest('blocks', { cursor });
   const dataCount = await getRequest('blocks/count');
+  if (data.message === 'Error') {
+    throw new Error(`Server Error : ${data.error}`);
+  }
+
   return <ListActions data={data} error={!data} totalCount={dataCount} />;
 };
 export default List;
